+++ app/pyfatoora/__init__.py
@@ -1,0 +1,1 @@
+from pyfatoora.pyfatoora import PyFatoora
