--- main.py
@@ -1,4 +1,0 @@
-from pyfatoora import app
-import uvicorn
-if __name__ == "__main__":
-    uvicorn.run(app)
