--- pyfatoora/__init__.py
@@ -1,2 +1,0 @@
-from .api import app
-from .pyfatoora import PyFatoora
